Move job states into parsl.jobs
