+++ parsl/executors/status_handling.py
@@ -8,7 +8,8 @@ from typing import List, Any, Dict, Optional, Tuple, Union
 import parsl  # noqa F401
 from parsl.executors.base import ParslExecutor
 from parsl.executors.errors import BadStateException, ScalingFailed
-from parsl.providers.base import JobStatus, ExecutionProvider, JobState
+from parsl.jobs.states import JobStatus, JobState
+from parsl.providers.base import ExecutionProvider
 from parsl.utils import AtomicIDCounter
