+++ parsl/jobs/job_error_handler.py
@@ -5,7 +5,7 @@ from typing import List, Dict
 import parsl.jobs.job_status_poller as jsp
 
 from parsl.executors.base import ParslExecutor
-from parsl.providers.base import JobStatus, JobState
+from parsl.jobs.states import JobStatus, JobState
 
 
 class JobErrorHandler:
