+++ parsl/providers/cobalt/cobalt.py
@@ -7,7 +7,7 @@ from parsl.channels import LocalChannel
 from parsl.launchers import AprunLauncher
 from parsl.providers.cobalt.template import template_string
 from parsl.providers.cluster_provider import ClusterProvider
-from parsl.providers.base import JobState, JobStatus
+from parsl.jobs.states import JobState, JobStatus
 from parsl.utils import RepresentationMixin, wtime_to_minutes
 
 logger = logging.getLogger(__name__)
