+++ parsl/providers/local/local.py
@@ -3,8 +3,9 @@ import os
 import time
 
 from parsl.channels import LocalChannel
+from parsl.jobs.states import JobState, JobStatus
 from parsl.launchers import SingleNodeLauncher
-from parsl.providers.base import ExecutionProvider, JobState, JobStatus
+from parsl.providers.base import ExecutionProvider
 from parsl.providers.errors import SchedulerMissingArgs, ScriptPathError, SubmitException
 from parsl.utils import RepresentationMixin
